Update mixins.md
Hi there,

Just checking that in the example in the "Good" section, the overlay_controller is empty - given that the `useOverlay` const has the relevant methods we are after? Otherwise things would be twice defined given the first `overlay_controller` (which also has those methods) and the const - which again has those same methods? hence this PR. feel free to close if i've made an error somewhere.

rgds

Ben